+++ source/extensions/omni.isaac.core/omni/isaac/core/objects/cuboid.py
@@ -20,22 +20,22 @@ from omni.isaac.core.utils.stage import get_current_stage
 
 
 class VisualCuboid(GeometryPrim):
-    """[summary]
+    """_summary_
 
     Args:
-        prim_path (str): [description]
-        name (str, optional): [description]. Defaults to "visual_cube".
-        position (Optional[Sequence[float]], optional): [description]. Defaults to None.
-        translation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-        orientation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-        scale (Optional[Sequence[float]], optional): [description]. Defaults to None.
-        visible (bool, optional): [description]. Defaults to True.
-        color (Optional[np.ndarray], optional): [description]. Defaults to None.
-        size (Optional[np.ndarray], optional): [description]. Defaults to None.
-        visual_material (Optional[VisualMaterial], optional): [description]. Defaults to None.
+        prim_path (str): _description_
+        name (str, optional): _description_. Defaults to "visual_cube".
+        position (Optional[Sequence[float]], optional): _description_. Defaults to None.
+        translation (Optional[Sequence[float]], optional): _description_. Defaults to None.
+        orientation (Optional[Sequence[float]], optional): _description_. Defaults to None.
+        scale (Optional[Sequence[float]], optional): _description_. Defaults to None.
+        visible (Optional[bool], optional): _description_. Defaults to None.
+        color (Optional[np.ndarray], optional): _description_. Defaults to None.
+        size (Optional[np.ndarray], optional): _description_. Defaults to None.
+        visual_material (Optional[VisualMaterial], optional): _description_. Defaults to None.
 
     Raises:
-        Exception: [description]
+        Exception: _description_
     """
 
     def __init__(
@@ -46,13 +46,11 @@ class VisualCuboid(GeometryPrim):
         translation: Optional[Sequence[float]] = None,
         orientation: Optional[Sequence[float]] = None,
         scale: Optional[Sequence[float]] = None,
-        visible: bool = True,
+        visible: Optional[bool] = None,
         color: Optional[np.ndarray] = None,
         size: Optional[np.ndarray] = None,
         visual_material: Optional[VisualMaterial] = None,
     ) -> None:
-        if size is None:
-            size = np.array([5, 5, 5])
         if is_prim_path_valid(prim_path):
             prim = get_prim_at_path(prim_path)
             if not prim.IsA(UsdGeom.Cube):
@@ -60,12 +58,17 @@ class VisualCuboid(GeometryPrim):
             cubeGeom = UsdGeom.Cube(prim)
         else:
             cubeGeom = UsdGeom.Cube.Define(get_current_stage(), prim_path)
-            cubeGeom.GetExtentAttr().Set(
-                [
-                    Gf.Vec3f([-size[0] / 2.0, -size[1] / 2.0, -size[2] / 2.0]),
-                    Gf.Vec3f([size[0] / 2.0, size[1] / 2.0, size[2] / 2.0]),
-                ]
-            )
+            if size is None:
+                size = np.array([0.05, 0.05, 0.05])
+            if visible is None:
+                visible = True
+            if visual_material is None:
+                if color is None:
+                    color = np.array([0.5, 0.5, 0.5])
+                visual_prim_path = find_unique_string_name(
+                    initial_name="/World/Looks/visual_material", is_unique_fn=lambda x: not is_prim_path_valid(x)
+                )
+                visual_material = PreviewSurface(prim_path=visual_prim_path, color=color)
         GeometryPrim.__init__(
             self,
             prim_path=prim_path,
@@ -77,16 +80,17 @@ class VisualCuboid(GeometryPrim):
             visible=visible,
             collision=False,
         )
-        VisualCuboid.set_size(self, size)
-        if not self.is_visual_material_applied():
-            if visual_material is None:
-                if color is None:
-                    color = np.array([0.5, 0.5, 0.5])
-                visual_prim_path = find_unique_string_name(
-                    initial_name="/World/Looks/visual_material", is_unique_fn=lambda x: not is_prim_path_valid(x)
-                )
-                visual_material = PreviewSurface(prim_path=visual_prim_path, color=color)
+        if visual_material is not None:
             VisualCuboid.apply_visual_material(self, visual_material)
+        if size is not None:
+            VisualCuboid.set_size(self, size)
+        size = VisualCuboid.get_size(self)
+        cubeGeom.GetExtentAttr().Set(
+            [
+                Gf.Vec3f([-size[0] / 2.0, -size[1] / 2.0, -size[2] / 2.0]),
+                Gf.Vec3f([size[0] / 2.0, size[1] / 2.0, size[2] / 2.0]),
+            ]
+        )
         return
 
     def set_size(self, size: np.ndarray) -> None:
@@ -105,49 +109,58 @@ class VisualCuboid(GeometryPrim):
         Returns:
             float: [description]
         """
-        return self.get_local_scale()
+        return self.get_local_scale() * self.geom.GetSizeAttr().Get()
 
 
 class FixedCuboid(VisualCuboid):
-    """[summary]
+    """_summary_
 
         Args:
-            prim_path (str): [description]
-            name (str, optional): [description]. Defaults to "dynamic_cube".
-            position (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            translation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            orientation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            scale (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            visible (bool, optional): [description]. Defaults to True.
-            color (Optional[np.ndarray], optional): [description]. Defaults to None.
-            static_friction (float, optional): [description]. Defaults to 0.2.
-            dynamic_friction (float, optional): [description]. Defaults to 1.0.
-            restitution (float, optional): [description]. Defaults to 0.0.
-            size (Optional[np.ndarray], optional): [description]. Defaults to None.
-            physics_material_path (Optional[PhysicsMaterial], optional): [description]. Defaults to None.
-            visual_material (Optional[VisualMaterial], optional): [description]. Defaults to None.
-
-        Raises:
-            Exception: [description]
+            prim_path (str): _description_
+            name (str, optional): _description_. Defaults to "fixed_cube".
+            position (Optional[np.ndarray], optional): _description_. Defaults to None.
+            translation (Optional[np.ndarray], optional): _description_. Defaults to None.
+            orientation (Optional[np.ndarray], optional): _description_. Defaults to None.
+            scale (Optional[np.ndarray], optional): _description_. Defaults to None.
+            visible (Optional[bool], optional): _description_. Defaults to None.
+            color (Optional[np.ndarray], optional): _description_. Defaults to None.
+            size (Optional[np.ndarray], optional): _description_. Defaults to None.
+            visual_material (Optional[VisualMaterial], optional): _description_. Defaults to None.
+            physics_material (Optional[PhysicsMaterial], optional): _description_. Defaults to None.
         """
 
     def __init__(
         self,
         prim_path: str,
-        name: str = "dynamic_cube",
+        name: str = "fixed_cube",
         position: Optional[np.ndarray] = None,
         translation: Optional[np.ndarray] = None,
         orientation: Optional[np.ndarray] = None,
         scale: Optional[np.ndarray] = None,
-        visible: bool = True,
+        visible: Optional[bool] = None,
         color: Optional[np.ndarray] = None,
-        static_friction: float = 0.2,
-        dynamic_friction: float = 1.0,
-        restitution: float = 0.0,
         size: Optional[np.ndarray] = None,
-        physics_material_path: Optional[PhysicsMaterial] = None,
         visual_material: Optional[VisualMaterial] = None,
+        physics_material: Optional[PhysicsMaterial] = None,
     ) -> None:
+        set_offsets = False
+        if not is_prim_path_valid(prim_path):
+            # set default values if no physics material given
+            if physics_material is None:
+                static_friction = 0.2
+                dynamic_friction = 1.0
+                restitution = 0.0
+                physics_material_path = find_unique_string_name(
+                    initial_name="/World/Physics_Materials/physics_material",
+                    is_unique_fn=lambda x: not is_prim_path_valid(x),
+                )
+                physics_material = PhysicsMaterial(
+                    prim_path=physics_material_path,
+                    dynamic_friction=dynamic_friction,
+                    static_friction=static_friction,
+                    restitution=restitution,
+                )
+            set_offsets = True
         VisualCuboid.__init__(
             self,
             prim_path=prim_path,
@@ -162,53 +175,35 @@ class FixedCuboid(VisualCuboid):
             visual_material=visual_material,
         )
         GeometryPrim.set_collision_enabled(self, True)
-
-        if physics_material_path is None:
-            physics_material_path = find_unique_string_name(
-                initial_name="/World/Physics_Materials/physics_material",
-                is_unique_fn=lambda x: not is_prim_path_valid(x),
-            )
-            my_physics_material = PhysicsMaterial(
-                prim_path=physics_material_path,
-                dynamic_friction=dynamic_friction,
-                static_friction=static_friction,
-                restitution=restitution,
-            )
-        else:
-            my_physics_material = PhysicsMaterial(prim_path=physics_material_path)
-
-        FixedCuboid.apply_physics_material(self, my_physics_material)
-        FixedCuboid.set_rest_offset(self, 0.0)
-        FixedCuboid.set_contact_offset(self, 0.1)
-        FixedCuboid.set_torsional_patch_radius(self, 1.0)
-        FixedCuboid.set_min_torsional_patch_radius(self, 0.8)
+        if physics_material is not None:
+            FixedCuboid.apply_physics_material(self, physics_material)
+        if set_offsets:
+            FixedCuboid.set_rest_offset(self, 0.0)
+            FixedCuboid.set_contact_offset(self, 0.1)
+            FixedCuboid.set_torsional_patch_radius(self, 1.0)
+            FixedCuboid.set_min_torsional_patch_radius(self, 0.8)
         return
 
 
 class DynamicCuboid(FixedCuboid, RigidPrim):
-    """[summary]
+    """_summary_
 
         Args:
-            prim_path (str): [description]
-            name (str, optional): [description]. Defaults to "dynamic_cube".
-            position (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            translation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            orientation (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            scale (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            visible (bool, optional): [description]. Defaults to True.
-            mass (Optional[float], optional): [description]. Defaults to 0.02.
-            color (Optional[np.ndarray], optional): [description]. Defaults to None.
-            linear_velocity (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            angular_velocity (Optional[Sequence[float]], optional): [description]. Defaults to None.
-            static_friction (float, optional): [description]. Defaults to 0.2.
-            dynamic_friction (float, optional): [description]. Defaults to 1.0.
-            restitution (float, optional): [description]. Defaults to 0.0.
-            size (Optional[np.ndarray], optional): [description]. Defaults to None.
-            physics_material_path (Optional[PhysicsMaterial], optional): [description]. Defaults to None.
-            visual_material (Optional[VisualMaterial], optional): [description]. Defaults to None.
-
-        Raises:
-            Exception: [description]
+            prim_path (str): _description_
+            name (str, optional): _description_. Defaults to "dynamic_cube".
+            position (Optional[np.ndarray], optional): _description_. Defaults to None.
+            translation (Optional[np.ndarray], optional): _description_. Defaults to None.
+            orientation (Optional[np.ndarray], optional): _description_. Defaults to None.
+            scale (Optional[np.ndarray], optional): _description_. Defaults to None.
+            visible (Optional[bool], optional): _description_. Defaults to None.
+            color (Optional[np.ndarray], optional): _description_. Defaults to None.
+            size (Optional[np.ndarray], optional): _description_. Defaults to None.
+            visual_material (Optional[VisualMaterial], optional): _description_. Defaults to None.
+            physics_material (Optional[PhysicsMaterial], optional): _description_. Defaults to None.
+            mass (Optional[float], optional): _description_. Defaults to None.
+            density (Optional[float], optional): _description_. Defaults to None.
+            linear_velocity (Optional[Sequence[float]], optional): _description_. Defaults to None.
+            angular_velocity (Optional[Sequence[float]], optional): _description_. Defaults to None.
         """
 
     def __init__(
@@ -219,19 +214,19 @@ class DynamicCuboid(FixedCuboid, RigidPrim):
         translation: Optional[np.ndarray] = None,
         orientation: Optional[np.ndarray] = None,
         scale: Optional[np.ndarray] = None,
-        visible: bool = True,
-        mass: Optional[float] = 0.02,
-        density: Optional[float] = None,
+        visible: Optional[bool] = None,
         color: Optional[np.ndarray] = None,
+        size: Optional[np.ndarray] = None,
+        visual_material: Optional[VisualMaterial] = None,
+        physics_material: Optional[PhysicsMaterial] = None,
+        mass: Optional[float] = None,
+        density: Optional[float] = None,
         linear_velocity: Optional[Sequence[float]] = None,
         angular_velocity: Optional[Sequence[float]] = None,
-        static_friction: float = 0.2,
-        dynamic_friction: float = 1.0,
-        restitution: float = 0.0,
-        size: Optional[np.ndarray] = None,
-        physics_material_path: Optional[PhysicsMaterial] = None,
-        visual_material: Optional[VisualMaterial] = None,
     ) -> None:
+        if not is_prim_path_valid(prim_path):
+            if mass is None:
+                mass = 0.02
         FixedCuboid.__init__(
             self,
             prim_path=prim_path,
@@ -242,12 +237,9 @@ class DynamicCuboid(FixedCuboid, RigidPrim):
             scale=scale,
             visible=visible,
             color=color,
-            static_friction=static_friction,
-            dynamic_friction=dynamic_friction,
-            restitution=restitution,
             size=size,
-            physics_material_path=physics_material_path,
             visual_material=visual_material,
+            physics_material=physics_material,
         )
         RigidPrim.__init__(
             self,