+++ source/standalone_examples/api/omni.isaac.kaya/kaya_move.py
@@ -41,13 +41,17 @@ my_world.scene.add_default_ground_plane()
 kaya_setup = HolonomicRobotUsdSetup(
     robot_prim_path=my_kaya.prim_path, com_prim_path="/World/Kaya/base_link/control_offset"
 )
-wheel_radius, wheel_positions, wheel_orientations, mecanum_angles = kaya_setup.get_holonomic_controller_params()
+wheel_radius, wheel_positions, wheel_orientations, mecanum_angles, wheel_axis, up_axis = (
+    kaya_setup.get_holonomic_controller_params()
+)
 my_controller = HolonomicController(
     name="holonomic_controller",
     wheel_radius=wheel_radius,
     wheel_positions=wheel_positions,
     wheel_orientations=wheel_orientations,
     mecanum_angles=mecanum_angles,
+    wheel_axis=wheel_axis,
+    up_axis=up_axis,
 )
 
 my_world.reset()